Add length to Buffer.concat invoke
Add length to Buffer.concat invoke.
That should increase performance according to the node documenation.

If totalLength is not provided, it is calculated from the Buffer instances in list. This however causes an additional loop to be executed in order to calculate the totalLength, so it is faster to provide the length explicitly if it is already known.